alles gefixed samen met jelco

+++ SomerenModel/Revenue.cs
@@ -10,6 +10,7 @@ namespace SomerenModel
     {
         public int TotalDrinksSold { get; set; }
         public int UniqueCustomers { get; set; }
-        public int Turnover { get; set; }
+        public double Turnover { get; set; }
+       
     }
 }
